security: apply least-privilege by removing chown on binaries
It is not needed for dogecoin:dogecoin to have permissions on
binaries installed in /usr/local/bin because we allow anyone to
execute these anyway. All this does is give running processes
like dogecoind access to these files, which the process then can
chmod and overwrite.

Instead, let root own the files, but keep allowing execution by
anyone, which will disallow a process spawned by the dogecoin
user to change these files.